move function ADR into POU adr
most situations were already covered by the POU adr
test that explains the {ref} macro is new and was moved
to the pou adr.
