+++ src/tests/adr.rs
@@ -1,6 +1,5 @@
 mod annotated_ast;
 /// tests for Architecture Design Records (ADR)
 mod enum_adr;
-mod function_calls_adr;
 mod pou_adr;
 mod util_macros;
